LIBFCREPO-1189. Coved common code out of the cli and import commands.
- created plastron.jobs.utils module
- added a configure_repo() method to the MessageProcessor
- extended the Repository.from_config() method to read all options from the config file
- moved tests to the appropriate packages
- moved model utility function to the plastron-models package

https://umd-dit.atlassian.net/browse/LIBFCREPO-1189
