+++ plastron-cli/tests/test_embedded_objects.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 
 import pytest
 
-from plastron.cli.commands.importcommand import create_repo_changeset
+from plastron.jobs.utils import create_repo_changeset
 from plastron.jobs import Row, build_fields
 from plastron.models import Item
