+++ plastron-models/src/plastron/models/__init__.py
@@ -1,4 +1,20 @@
-from .letter import Letter
-from .newspaper import Issue
-from .poster import Poster
-from .umd import Item, umdtype, umdform
+from typing import Type
+
+from plastron.models.letter import Letter
+from plastron.models.newspaper import Issue
+from plastron.models.poster import Poster
+from plastron.models.umd import Item, umdtype, umdform
+from plastron.rdf.rdf import Resource
+
+
+class ModelClassNotFoundError(Exception):
+    def __init__(self, model_name: str, *args):
+        super().__init__(*args)
+        self.model_name = model_name
+
+
+def get_model_class(model_name: str) -> Type[Resource]:
+    try:
+        return getattr(__package__, model_name)
+    except AttributeError as e:
+        raise ModelClassNotFoundError(model_name) from e
