+++ plastron-repo/src/plastron/jobs/__init__.py
@@ -3,7 +3,6 @@ import logging
 import os
 import re
 import urllib.parse
-from collections import defaultdict
 from dataclasses import dataclass
 from datetime import datetime
 from email.utils import parsedate
@@ -13,7 +12,7 @@ from pathlib import Path
 from shutil import copyfileobj
 from tempfile import TemporaryDirectory
 from time import mktime
-from typing import List, Mapping, Optional
+from typing import List, Mapping, Optional, Type
 from urllib.parse import urlsplit
 from zipfile import ZipFile
 
@@ -21,123 +20,22 @@ import yaml
 from bagit import make_bag
 from paramiko import SFTPClient, SSHException
 from rdflib import URIRef
-from rdflib.util import from_n3
 from requests import ConnectionError
 
-import plastron.models
 from plastron.client import Client, ClientError
 from plastron.core.exceptions import DataReadException, FailureException
-from plastron.core.util import ItemLog, datetimestamp
-from plastron.files import get_ssh_client
-from plastron.models import Item
-from plastron.namespaces import get_manager
-from plastron.rdf.pcdm import File
-from plastron.rdf.rdf import RDFDataProperty, Resource
-from plastron.serializers import CSVSerializer, SERIALIZER_CLASSES, detect_resource_class, EmptyItemListError
+from plastron.core.util import ItemLog, datetimestamp, strtobool
+from plastron.files import get_ssh_client, ZipFileSource, RemoteFileSource, HTTPFileSource, LocalFileSource
+from plastron.jobs.utils import create_repo_changeset, build_file_groups, annotate_from_files, build_fields
+from plastron.models import get_model_class, Item, ModelClassNotFoundError
+from plastron.rdf.pcdm import File, PreservationMasterFile
+from plastron.rdf.rdf import Resource
+from plastron.repo import Repository
+from plastron.serializers import SERIALIZER_CLASSES, detect_resource_class, EmptyItemListError
+from plastron.validation import validate, ValidationError
 
-nsm = get_manager()
 logger = logging.getLogger(__name__)
 UUID_REGEX = re.compile(r'([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})', re.IGNORECASE)
-
-
-class ModelClassNotFoundError(Exception):
-    def __init__(self, model_name: str, *args):
-        super().__init__(*args)
-        self.model_name = model_name
-
-
-def build_lookup_index(item: Resource, index_string: str):
-    """
-    Build a lookup dictionary for embedded object properties of an item.
-
-    :param item:
-    :param index_string:
-    :return:
-    """
-    index = defaultdict(dict)
-    if index_string is None:
-        return index
-
-    pattern = r'([\w]+)\[(\d+)\]'
-    for entry in index_string.split(';'):
-        key, uriref = entry.split('=')
-        m = re.search(pattern, key)
-        attr = m[1]
-        i = int(m[2])
-        prop = getattr(item, attr)
-        try:
-            index[attr][i] = prop[URIRef(item.uri + uriref)]
-        except IndexError:
-            # need to create an object with that URI
-            obj = prop.obj_class(uri=URIRef(item.uri + uriref))
-            # TODO: what if i > 0?
-            prop.values.append(obj)
-            index[attr][i] = obj
-    return index
-
-
-def build_fields(fieldnames, model_class):
-    property_attrs = {header: attrs for attrs, header in model_class.HEADER_MAP.items()}
-    fields = defaultdict(list)
-    # group typed and language-tagged columns by their property attribute
-    for header in fieldnames:
-        if '[' in header:
-            # this field has a language tag
-            # header format is "Header Label [Language Label]"
-            header_label, language_label = re.search(r'^([^[]+)\s+\[(.+)]$', header).groups()
-            try:
-                attrs = property_attrs[header_label]
-            except KeyError as e:
-                raise DataReadException(f'Unknown header "{header}" in import file.') from e
-            # if the language label isn't a name in the LANGUAGE_CODES table,
-            # assume that it is itself a language code
-            lang_code = CSVSerializer.LANGUAGE_CODES.get(language_label, language_label)
-            fields[attrs].append({
-                'header': header,
-                'lang_code': lang_code,
-                'datatype': None
-            })
-        elif '{' in header:
-            # this field has a datatype
-            # header format is "Header Label {Datatype Label}
-            header_label, datatype_label = re.search(r'^([^{]+)\s+{(.+)}$', header).groups()
-            try:
-                attrs = property_attrs[header_label]
-            except KeyError as e:
-                raise DataReadException(f'Unknown header "{header}" in import file.') from e
-            # the datatype label should either be a key in the lookup table,
-            # or an n3-abbreviated URI of a datatype
-            try:
-                datatype_uri = CSVSerializer.DATATYPE_URIS.get(datatype_label, from_n3(datatype_label, nsm=nsm))
-                if not isinstance(datatype_uri, URIRef):
-                    raise DataReadException(f'Unknown datatype "{datatype_label}" in "{header}" in import file.')
-            except KeyError as e:
-                raise DataReadException(f'Unknown datatype "{datatype_label}" in "{header}" in import file.') from e
-
-            fields[attrs].append({
-                'header': header,
-                'lang_code': None,
-                'datatype': datatype_uri
-            })
-        else:
-            # no language tag or datatype
-            # make sure we skip the system columns
-            if header not in CSVSerializer.SYSTEM_HEADERS:
-                if header not in property_attrs:
-                    raise DataReadException(f'Unrecognized header "{header}" in import file.')
-                # check for a default datatype defined in the model
-                attrs = property_attrs[header]
-                prop = model_class.name_to_prop.get(attrs)
-                if prop is not None and issubclass(prop, RDFDataProperty):
-                    datatype_uri = prop.datatype
-                else:
-                    datatype_uri = None
-                fields[attrs].append({
-                    'header': header,
-                    'lang_code': None,
-                    'datatype': datatype_uri
-                })
-    return fields
 
 
 def is_run_dir(path: Path) -> bool:
@@ -228,10 +126,7 @@ class ImportJob:
     @property
     def model_class(self):
         if self._model_class is None:
-            try:
-                self._model_class = getattr(plastron.models, self.model)
-            except AttributeError as e:
-                raise ModelClassNotFoundError(self.model) from e
+            self._model_class = get_model_class(self.model)
         return self._model_class
 
     def complete(self, item: Resource, line_reference: str, status: ImportedItemStatus):
@@ -276,6 +171,7 @@ class ImportRun:
     """
     A single run of an import job. Records the logs of invalid and failed items (if any).
     """
+
     def __init__(self, job: ImportJob):
         self.job = job
         self.dir = None
@@ -383,6 +279,7 @@ class MetadataRows:
     """
     Iterable sequence of rows from the metadata CSV file of an import job.
     """
+
     def __init__(self, job: ImportJob, limit: int = None, percentage: int = None):
         self.job = job
         self.limit = limit
@@ -558,6 +455,328 @@ class Row:
     @property
     def index_string(self):
         return self.data.get('INDEX')
+
+
+class ImportOperation:
+    def __init__(self, job_id: str, jobs_dir: str, repo: Repository, ssh_private_key: str = None):
+        self.job = ImportJob(job_id=job_id, jobs_dir=jobs_dir)
+        logger.debug(f'Job directory is {self.job.dir}')
+        self.repo = repo
+        self.ssh_private_key = ssh_private_key
+
+    def start(
+            self,
+            model: Type[Resource],
+            access: URIRef = None,
+            member_of: URIRef = None,
+            container: str = None,
+            binaries_location: str = None,
+            **kwargs,
+    ):
+        self.job.save_config({
+            'model': model,
+            'access': access,
+            'member_of': member_of,
+            'container': container,
+            'binaries_location': binaries_location
+        })
+
+        return (yield from self.import_items(**kwargs))
+
+    def resume(self, **kwargs):
+        if not self.job.dir_exists:
+            raise RuntimeError(f'Cannot resume job "{self.job.id}": no such job directory: "{self.job.dir}"')
+
+        logger.info(f'Resuming saved job {self.job.id}')
+        try:
+            # load stored config from the previous run of this job
+            self.job.load_config()
+        except FileNotFoundError:
+            raise RuntimeError(f'Cannot resume job {self.job.id}: no config.yml found in {self.job.dir}')
+
+        return (yield from self.import_items(**kwargs))
+
+    def import_items(self, limit: int = None, percentage=None, validate_only: bool = False, import_file=None):
+        start_time = datetime.now().timestamp()
+
+        if percentage:
+            logger.info(f'Loading {percentage}% of the total items')
+        if validate_only:
+            logger.info('Validation-only mode, skipping imports')
+
+        # if an import file was provided, save that as the new CSV metadata file
+        if import_file is not None:
+            self.job.store_metadata_file(import_file)
+
+        try:
+            metadata = self.job.metadata(limit=limit, percentage=percentage)
+        except ModelClassNotFoundError as e:
+            raise FailureException(f'Model class {e.model_name} not found') from e
+        except JobError as e:
+            raise FailureException(str(e)) from e
+
+        if metadata.has_binaries and self.job.binaries_location is None:
+            raise RuntimeError('Must specify --binaries-location if the metadata has a FILES column')
+
+        initial_completed_item_count = len(self.job.completed_log)
+        logger.info(f'Found {initial_completed_item_count} completed items')
+
+        import_run = self.job.new_run().start()
+        for row in metadata:
+            repo_changeset = create_repo_changeset(self.repo.client, metadata, row)
+            item = repo_changeset.item
+
+            # count the number of files referenced in this row
+            metadata.files += len(row.filenames)
+
+            try:
+                report = validate(item)
+            except ValidationError as e:
+                raise RuntimeError(f'Unable to run validation: {e}') from e
+
+            metadata.validation_reports.append({
+                'line': row.line_reference,
+                'is_valid': report.is_valid(),
+                'passed': [outcome for outcome in report.passed()],
+                'failed': [outcome for outcome in report.failed()]
+            })
+
+            missing_files = [
+                name for name in row.filenames if not self.get_source(self.job.binaries_location, name).exists()
+            ]
+            if len(missing_files) > 0:
+                logger.warning(f'{len(missing_files)} file(s) for "{item}" not found')
+
+            if report.is_valid() and len(missing_files) == 0:
+                metadata.valid += 1
+                logger.info(f'"{item}" is valid')
+            else:
+                # drop invalid items
+                metadata.invalid += 1
+                logger.warning(f'"{item}" is invalid, skipping')
+                reasons = [' '.join(str(f) for f in outcome) for outcome in report.failed()]
+                if len(missing_files) > 0:
+                    reasons.extend(f'Missing file: {f}' for f in missing_files)
+                import_run.drop_invalid(
+                    item=item,
+                    line_reference=row.line_reference,
+                    reason=f'Validation failures: {"; ".join(reasons)}'
+                )
+                continue
+
+            if validate_only:
+                # validation-only mode
+                continue
+
+            try:
+                self.update_repo(metadata, row, repo_changeset)
+            except RuntimeError as e:
+                metadata.errors += 1
+                logger.error(f'{item} import failed: {e}')
+                import_run.drop_failed(item, row.line_reference, reason=str(e))
+
+            # update the status
+            now = datetime.now().timestamp()
+            yield {
+                'time': {
+                    'started': start_time,
+                    'now': now,
+                    'elapsed': now - start_time
+                },
+                'count': metadata.stats()
+            }
+
+        return metadata
+
+    def update_repo(self, metadata, row, repo_changeset):
+        """
+        Updates the repository with the given RepoChangeSet
+
+        :param metadata: A plastron.jobs.MetadataRows object representing the
+                          CSV file being imported
+        :param row: A single plastron.jobs.Row object representing the row
+                     being imported
+        :param repo_changeset: The RepoChangeSet object describing the changes
+                                 to make to the repository.
+        """
+        created_uris = []
+        updated_uris = []
+        item = repo_changeset.item
+
+        if not item.created:
+            # if an item is new, don't construct a SPARQL Update query
+            # instead, just create and update normally
+            # create new item in the repo
+            logger.debug('Creating a new item')
+            # add the access class
+            if self.job.access is not None:
+                item.rdf_type.append(URIRef(self.job.access))
+            # add the collection membership
+            if self.job.member_of is not None:
+                item.member_of = URIRef(self.job.member_of)
+
+            if row.has_files:
+                create_pages = bool(strtobool(row.get('CREATE_PAGES', 'True')))
+                logger.debug('Adding pages and files to new item')
+                self.add_files(
+                    item,
+                    build_file_groups(row['FILES']),
+                    base_location=self.job.binaries_location,
+                    access=self.job.access,
+                    create_pages=create_pages
+                )
+
+            if row.has_item_files:
+                self.add_files(
+                    item,
+                    build_file_groups(row['ITEM_FILES']),
+                    base_location=self.job.binaries_location,
+                    access=self.job.access,
+                    create_pages=False
+                )
+
+            if args.extract_text_types is not None:
+                annotate_from_files(item, args.extract_text_types.split(','))
+
+            logger.debug(f"Creating resources in container: {self.job.container}")
+
+            try:
+                with self.repo.client.transaction() as txn_client:
+                    item.create(txn_client, container_path=self.job.container)
+                    item.update(txn_client)
+                    txn_client.commit()
+            except Exception as e:
+                raise RuntimeError(f'Creating item failed: {e}') from e
+
+            self.job.complete(item, row.line_reference, ImportedItemStatus.CREATED)
+            metadata.created += 1
+            created_uris.append(item.uri)
+
+        elif repo_changeset:
+            # construct the SPARQL Update query if there are any deletions or insertions
+            # then do a PATCH update of an existing item
+            logger.info(f'Sending update for {item}')
+            sparql_update = repo_changeset.build_sparql_update(self.repo.client)
+            logger.debug(sparql_update)
+            try:
+                item.patch(self.repo.client, sparql_update)
+            except ClientError as e:
+                raise RuntimeError(f'Updating item failed: {e}') from e
+
+            self.job.complete(item, row.line_reference, ImportedItemStatus.MODIFIED)
+            metadata.updated += 1
+            updated_uris.append(item.uri)
+
+        else:
+            self.job.complete(item, row.line_reference, ImportedItemStatus.UNCHANGED)
+            metadata.unchanged += 1
+            logger.info(f'No changes found for "{item}" ({row.uri}); skipping')
+            metadata.skipped += 1
+
+    def get_source(self, base_location, path):
+        """
+        Get an appropriate BinarySource based on the type of ``base_location``.
+        The following forms of ``base_location`` are recognized:
+
+        * ``zip:<path to zipfile>``
+        * ``sftp:<user>@<host>/<path to dir>``
+        * ``http://<host>/<path to dir>``
+        * ``zip+sftp:<user>@<host>/<path to zipfile>``
+        * ``<local dir path>``
+
+        :param base_location:
+        :param path:
+        :return:
+        """
+        if base_location.startswith('zip:'):
+            return ZipFileSource(base_location[4:], path)
+        elif base_location.startswith('sftp:'):
+            return RemoteFileSource(
+                location=os.path.join(base_location, path),
+                ssh_options={'key_filename': self.ssh_private_key}
+            )
+        elif base_location.startswith('http:') or base_location.startswith('https:'):
+            base_uri = base_location if base_location.endswith('/') else base_location + '/'
+            return HTTPFileSource(base_uri + path)
+        elif base_location.startswith('zip+sftp:'):
+            return ZipFileSource(
+                zip_file=base_location[4:],
+                path=path,
+                ssh_options={'key_filename': self.ssh_private_key}
+            )
+        else:
+            # with no URI prefix, assume a local file path
+            return LocalFileSource(localpath=os.path.join(base_location, path))
+
+    def get_file(self, base_location, filename):
+        """
+        Get a file object for the given base_location and filename.
+
+        Currently, if the file has an "image/tiff" MIME type, this method returns
+        a :py:class:`plastron.pcdm.PreservationMasterFile`; otherwise it returns
+        a basic :py:class:`plastron.pcdm.File`.
+
+        :param base_location:
+        :param filename:
+        :return:
+        """
+        source = self.get_source(base_location, filename)
+
+        # XXX: hardcoded image/tiff as the preservation master format
+        # TODO: make preservation master format configurable per collection or job
+        if source.mimetype() == 'image/tiff':
+            file_class = PreservationMasterFile
+        else:
+            file_class = File
+
+        return file_class.from_source(title=basename(filename), source=source)
+
+    def add_files(self, item, file_groups, base_location, access=None, create_pages=True):
+        """
+        Add pages and files to the given item. A page is added for each key (basename) in the file_groups
+        parameter, and a file is added for each element in the value list for that key.
+
+        :param item: PCDM Object to add the pages to.
+        :param file_groups: Dictionary of basename to filename list mappings.
+        :param base_location: Location of the files.
+        :param access: Optional RDF class representing the access level for this item.
+        :param create_pages: Whether to create an intermediate page object for each file group. Defaults to True.
+        :return: The number of files added.
+        """
+        if base_location is None:
+            raise RuntimeError('Must specify a binaries-location')
+
+        if create_pages:
+            logger.debug(f'Creating {len(file_groups.keys())} page(s)')
+
+        count = 0
+
+        for n, (rootname, filenames) in enumerate(file_groups.items(), 1):
+            if create_pages:
+                # create a member object for each rootname
+                # delegate to the item model how to build the member object
+                member = item.get_new_member(rootname, n)
+                # add to the item
+                item.add_member(member)
+                proxy = item.append_proxy(member, title=member.title)
+                # add the access class to the member resources
+                if access is not None:
+                    member.rdf_type.append(access)
+                    proxy.rdf_type.append(access)
+                file_parent = member
+            else:
+                # files will be added directly to the item
+                file_parent = item
+
+            # add the files to their parent object (either the item or a member)
+            for filename in filenames:
+                file = self.get_file(base_location, filename)
+                count += 1
+                file_parent.add_file(file)
+                if access is not None:
+                    file.rdf_type.append(access)
+
+        return count
 
 
 @dataclass
